requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
bootstrap4==0.1.0
cffi==1.14.5
chromelogger==0.4.3
colorama==0.4.4
Django==3.1.7
django-bootstrap3==14.2.0
django-bootstrap4==3.0.1
django-braces==1.14.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-developer-panel==0.1.2
django-misaka==0.2.1
django-widget-tweaks==1.4.8
gunicorn==20.1.0
houdini.py==0.1.0
isort==5.7.0
jedi==0.18.0
jsonpickle==2.0.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
misaka==2.1.1
parso==0.8.1
Pillow==8.2.0
pycparser==2.20
Pygments==2.9.0
pylint==2.6.0
pytz==2021.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
